Assert available for send assets
