+++ packages/app/src/ui/molecules/asset-selector/AssetSelector.tsx
@@ -34,13 +34,13 @@ export function AssetSelector({
     return (
       <div
         className="flex h-14 w-36 flex-row items-center justify-center rounded-xl border border-input bg-input-background ring-offset-background"
-        data-testid={testIds.component.AssetSelector}
+        data-testid={testIds.component.AssetSelector.trigger}
       >
         {selectedAsset ? (
-          <>
+          <div className="flex" data-testid={testIds.component.AssetSelector.option}>
             <TokenIcon token={selectedAsset} className="h-6 w-6" />
             <Typography className="ml-2 text-prompt-foreground">{selectedAsset.symbol}</Typography>
-          </>
+          </div>
         ) : (
           '-'
         )}
@@ -50,7 +50,7 @@ export function AssetSelector({
 
   return (
     <Select open={open} value={selectedAsset?.symbol} onValueChange={setSelectedAsset}>
-      <SelectTrigger disabled={disabled} className="h-14 w-36" data-testid={testIds.component.AssetSelector}>
+      <SelectTrigger disabled={disabled} className="h-14 w-36" data-testid={testIds.component.AssetSelector.trigger}>
         <Wrapper>
           <div className="flex flex-row items-center gap-2">
             {selectedAsset && <TokenIcon token={selectedAsset} className="h-6 w-6" />}
@@ -62,7 +62,7 @@ export function AssetSelector({
         {assets.map((a) => (
           <SelectItem key={a.token.symbol} value={a.token.symbol}>
             <div className={cn('flex flex-row justify-between', a.balance && 'w-40')}>
-              <div className="flex flex-row gap-2">
+              <div className="flex flex-row gap-2" data-testid={testIds.component.AssetSelector.option}>
                 <TokenIcon token={a.token} className="h-6 w-6" />
                 <Typography className="text-prompt-foreground">{a.token.symbol}</Typography>
               </div>
